add seperate test for trend calcs
